+++ src/programs/fs.ts
@@ -192,7 +192,7 @@ export const FS_CREATE_FILE: ProgramInfo = {
 
         [0x90, Opcode.SYSCALL],
         [0x91, 0],               // ← Syscall 0 = exit
-    ] as [u8, u8][]),
+    ] as [u16, u8][]),
 };
 
 
@@ -378,7 +378,7 @@ export const FS_CREATE_FILE_BIS: ProgramInfo = {
 
         [0x90, Opcode.SYSCALL],
         [0x91, 0],               // ← Syscall 0 = exit
-    ] as [u8, u8][]),
+    ] as [u16, u8][]),
 };
 
 
@@ -472,7 +472,7 @@ export const FS_READ_FILE: ProgramInfo = {
 
         [0x41, Opcode.SYSCALL],
         [0x42, 0],               // ← Syscall 0 = exit
-    ] as [u8, u8][]),
+    ] as [u16, u8][]),
 };
 
 
@@ -561,7 +561,7 @@ export const FS_READ_FILE_BIS: ProgramInfo = {
 
         [0x41, Opcode.SYSCALL],
         [0x42, 0],               // ← Syscall 0 = exit
-    ] as [u8, u8][]),
+    ] as [u16, u8][]),
 };
 
 
@@ -590,7 +590,7 @@ export const FS_LIST_FILES: ProgramInfo = {
 
         [0x09, Opcode.SYSCALL],
         [0x0A, 0],               // ← Syscall 0 = exit
-    ] as [u8, u8][]),
+    ] as [u16, u8][]),
 };
