+++ src/programs/timer.ts
@@ -2,7 +2,7 @@
 import { Opcode } from "@/lib/instructions";
 import { MEMORY_MAP } from "@/lib/memory_map";
 
-import type { ProgramInfo, u8 } from "@/types/cpu.types";
+import type { ProgramInfo, u16, u8 } from "@/types/cpu.types";
 
 
 export const programs: Record<string, ProgramInfo> = {
@@ -80,7 +80,7 @@ export const programs: Record<string, ProgramInfo> = {
 
             [0x50, Opcode.POP_A],
             [0x51, Opcode.IRET],
-        ] as [u8, u8][]),
+        ] as [u16, u8][]),
     },
 
     timer_02: {
@@ -151,7 +151,7 @@ export const programs: Record<string, ProgramInfo> = {
 
             [0x60, Opcode.POP_A],
             [0x61, Opcode.IRET], // Retourne à 0x0229 (le HALT)
-        ] as [u8, u8][]),
+        ] as [u16, u8][]),
     },
 }
